# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PEGAPROMO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register entrepreneur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /collaborators:
    get:
      operationId: listCollaborators
      tags: [collaborators]
      summary: List collaborators
      responses:
        "200":
          description: Collaborator list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collaborator"
    post:
      operationId: inviteCollaborator
      tags: [collaborators]
      summary: Invite collaborator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollaboratorInvite"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collaborator"
  /collaborators/{id}:
    put:
      operationId: updateCollaborator
      tags: [collaborators]
      summary: Update collaborator permissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollaboratorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collaborator"
    delete:
      operationId: removeCollaborator
      tags: [collaborators]
      summary: Remove collaborator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        planId:
          type: integer
      required:
        - name
        - email
        - password
        - phone
        - planId
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, entrepreneur, collaborator]
        planId:
          type: integer
        trialEndsAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [trial, active, overdue, cancelled]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - status
...
    Collaborator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, pending, removed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - permissions
        - status
    CollaboratorInvite:
      type: object
      properties:
        email:
          type: string
        permissions:
          type: array
          items:
            type: string
      required:
        - email
        - permissions
    CollaboratorUpdate:
      type: object
      properties:
        permissions:
          type: array
          items:
            type: string
      required:
        - permissions
...