# lib/api-zod/src/generated/types/collaboratorInvite.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PEGAPROMO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CollaboratorInvite {
  email: string;
  permissions: string[];
}